Completely refactor conferences events form
The form now uses the stimulus repeater controller inclued with Spina.
As a result of this, the custom stimulus controller made for this form
has been deleted. The events controller and its related routes have also
been removed, as the form no longer makes an AJAX request for the new
events form, instead including this within the form as per how the
repeater part works within Spina.